// zip_extracted/lib/types.ts
export interface TratamientoLente {
  diseno: 'monofocal' | 'bifocal' | 'progresivo' | 'ocupacional';
  fabricacion: 'terminado' | 'tallado-convencional' | 'freeform';
  material: 'cr39' | 'policarbonato' | 'trivex' | '1.60' | '1.67' | '1.74';
  antirreflejo: string; // ID o nombre del tratamiento AR
  fotocromatico: 'ninguno' | 'transitions-gen8' | 'transitions-xtractive' | 'sensity' | 'estandar';
  fotocromaticoColor?: 'gris' | 'cafe' | 'verde' | 'ambar';
  blueBlock: boolean;
  blueBlockTipo?: string;
}